requirements.txt
# 🤖 AI/ML Libraries
openai>=1.3.0
google-generativeai==0.7.2
ollama==0.3.3

# 🗄️ Database Connectors
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# 🌐 Web Framework
streamlit>=1.28.0

# 📊 Data Processing
pandas>=1.5.0
numpy>=1.24.0

# 🔧 Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# 🔒 Security
cryptography>=41.0.0
bcrypt>=4.0.0

# 🛠️ Development & Utilities
requests>=2.31.0
urllib3>=2.0.0

# 📝 Optional: Development Tools (remove in production)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
